+++ paradox-language-support/src/main/kotlin/com/windea/plugin/idea/paradox/settings/ParadoxSettingsComponent.kt
@@ -22,60 +22,11 @@ import javax.swing.*
 class ParadoxSettingsComponent {
 	val resolveInternalReferencesCheckBox = JBCheckBox(message("paradox.settings.resolveInternalReferences"))
 	val resolveExternalReferencesCheckBox = JBCheckBox(message("paradox.settings.resolveExternalReferences"))
-
-	//val useSteamDirectoryCheckBox = JBCheckBox(message("paradox.settings.useSteamDirectory")).apply {
-	//	addChangeListener {
-	//		//it: JBCheckBox
-	//		val useSteamDirectory = (it.source as JBCheckBox).isSelected
-	//		steamDirectoryTextField.isEnabled = useSteamDirectory
-	//		paradoxDirectoryTextField.isEnabled = !useSteamDirectory
-	//		paradoxModsDirectoryTextField.isEnabled = !useSteamDirectory
-	//	}
-	//}
-	//private val steamDirectoryLabel = JBLabel(message("paradox.settings.steamDirectory"))
-	//private val steamDirectoryTextFieldBacking = JBTextField()
-	//val steamDirectoryTextField = TextFieldWithBrowseButton(steamDirectoryTextFieldBacking)
-	private val stellarisDirectoryLabel = JBLabel(message("paradox.settings.paradoxDirectory"))
-	private val stellarisDirectoryTextFieldBacking = JBTextField()
-	val stellarisDirectoryTextField = TextFieldWithBrowseButton(stellarisDirectoryTextFieldBacking)
-	//private val paradoxModsDirectoryLabel = JBLabel(message("paradox.settings.paradoxMods.Directory"))
-	//private val paradoxModsDirectoryTextFieldBacking = JBTextField()
-	//val paradoxModsDirectoryTextField = TextFieldWithBrowseButton(paradoxModsDirectoryTextFieldBacking)
-
-	private val pathsPanel = FormBuilder.createFormBuilder()
-		//.addComponent(useSteamDirectoryCheckBox, 1)
-		//.addLabeledComponent(steamDirectoryLabel, steamDirectoryTextField, 1, false)
-		.addLabeledComponent(stellarisDirectoryLabel, stellarisDirectoryTextField, 1, false)
-		//.addLabeledComponent(paradoxModsDirectoryLabel, paradoxModsDirectoryTextField, 1, false)
-		.panel.apply {
-			border = IdeBorderFactory.createTitledBorder(message("paradox.settings.paths"))
-		}
-
+	
 	val panel = FormBuilder.createFormBuilder()
 		.addComponent(resolveInternalReferencesCheckBox,1)
 		.addComponent(resolveExternalReferencesCheckBox,1)
-		.addComponent(pathsPanel)
 		.addComponentFillVertically(JPanel(), 0)
 		.panel
-
-	init {
-		//SwingHelper.installFileCompletionAndBrowseDialog(
-		//	null,
-		//	steamDirectoryTextField,
-		//	message("paradox.settings.steamDirectory.select"),
-		//	FileChooserDescriptorFactory.createSingleFileDescriptor()
-		//)
-		SwingHelper.installFileCompletionAndBrowseDialog(
-			null,
-			stellarisDirectoryTextField,
-			message("paradox.settings.paradoxDirectory.select"),
-			FileChooserDescriptorFactory.createSingleFileDescriptor()
-		)
-		//SwingHelper.installFileCompletionAndBrowseDialog(
-		//	null,
-		//	paradoxModsDirectoryTextField,
-		//	message("paradox.settings.paradoxModsDirectory.select"),
-		//	FileChooserDescriptorFactory.createSingleFileDescriptor()
-		//)
-	}
+	
 }